Time: 0 ms (100%), Space: 25.6 MB (59.55%) - LeetHub

+++ 3194-minimum-average-of-smallest-and-largest-elements/3194-minimum-average-of-smallest-and-largest-elements.cpp
@@ -7,7 +7,7 @@ public:
         int i = 0;
         int j = n-1;
         while(i<j){
-            float a = (nums[i]+nums[j])/2.0;
+            float a = (float)(nums[i]+nums[j])/2;
             v.push_back(a);
             i++;
             j--;
